Y: makefile struggle is real

+++ lib/seccomp/rules.h
@@ -5,7 +5,7 @@
 #include <sys/syscall.h>
 
 #define ALLOW_SYSCALL(name) \
-    BPF_JUMP(BPF_JMP | BPF_JEQ | BPF_K, SYS_##name, 0, 1), \
+    BPF_JUMP(BPF_JMP | BPF_JEQ | BPF_K, (unsigned int)SYS_##name, 0, 1), \
     BPF_STMT(BPF_RET | BPF_K, SECCOMP_RET_ALLOW)
 
 #define KILL_PROCESS \
